+++ src/main/java/com/flippingcopilot/model/SavedOffer.java
@@ -12,8 +12,8 @@ public class SavedOffer
 	private int itemId;
 	private int quantitySold;
 	private int totalQuantity;
-	private int price;
-	private int spent;
+	private long price;
+	private long spent;
 	private GrandExchangeOfferState state;
 	private boolean copilotPriceUsed;
 	private boolean wasCopilotSuggestion;
